rename video_mode -> video_mode_app
