+++ src/qua_dashboards/pages/data_dashboard.py
@@ -1,6 +1,6 @@
 import dash
 from qua_dashboards.logging_config import logger
-from qua_dashboards.data_dashboard.app import DataDashboardApp
+from qua_dashboards.data_dashboard.data_dashboard_app import DataDashboardApp
 
 
 logger.info("Registering page data-dashboard")
